requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==5.1.4
djangorestframework==3.15.2
drf-spectacular==0.28.0
exceptiongroup==1.2.2
frozenlist==1.5.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
msgpack==1.1.0
multidict==6.1.0
packaging==24.2
pluggy==1.5.0
propcache==0.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
pytest==8.3.4
pytest-asyncio==0.24.0
python-decouple==3.8
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
rpds-py==0.22.3
service-identity==24.2.0
sqlparse==0.5.2
tomli==2.2.1
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
yarl==1.18.3
zope.interface==7.2
psycopg==3.2.3
python-decouple==3.8
